+++ CHANGELOG.md
@@ -3,16 +3,19 @@
 ## Changelog
 
 ### Version 1.2.0 (16-Oct-2025)
+
 > :warning: **Breaking Change** :warning:
 >
-> For users of version `1.0.0`, please first uninstall version `1.0.0` **before** installing any later version via:
+> For users of version `1.0.0` _only_, please first uninstall version `1.0.0` **before** installing any later version via:
 > 
 > `resetConfiguration="${4:-"Uninstall"}"`
 >
 > Please feel free to reach out to the Mac Admins Slack [#ddm-os-reminders](https://slack.com/app_redirect?channel=C09LVE2NVML) channel for assistance.
 > 
 > _Sorry for any Dan-induced headaches._
+
 - Addressed Issue #3: Use Dynamic icon based on OS Update version (thanks for the suggestion, @ScottEKendall!)
+- Addressed Issue #5: Added logic to ignore Do Not Disturb / Display Assertions 24 hours prior to enforcement (per [Apple's documentation](https://support.apple.com/guide/deployment/install-and-enforce-software-updates-depd30715cbb/1/web/1.0))
 
 ### Version 1.1.0 (16-Oct-2025)
 - Added `checkUserFocusDisplayAssertions` function to avoid interrupting users with Focus modes or Display Sleep Assertions enabled (thanks, @TechTrekkie!)