+++ CHANGELOG.md
@@ -15,10 +15,10 @@
 > _Sorry for any Dan-induced headaches._
 
 - Addressed Issue #3: Use Dynamic icon based on OS Update version (thanks for the suggestion, @ScottEKendall!)
-- Addressed Issue #5: Added logic to ignore Do Not Disturb / Display Assertions 24 hours prior to enforcement (per [Apple's documentation](https://support.apple.com/guide/deployment/install-and-enforce-software-updates-depd30715cbb/1/web/1.0))
+- Addressed Issue #5: Added logic to ignore Display Assertions 24 hours prior to enforcement (per [Apple's documentation](https://support.apple.com/guide/deployment/install-and-enforce-software-updates-depd30715cbb/1/web/1.0))
 
 ### Version 1.1.0 (16-Oct-2025)
-- Added `checkUserFocusDisplayAssertions` function to avoid interrupting users with Focus modes or Display Sleep Assertions enabled (thanks, @TechTrekkie!)
+- Added `checkUserFocusDisplayAssertions` function to avoid interrupting users with Display Sleep Assertions enabled (thanks, @TechTrekkie!)
 - Refactored `infobuttonaction` to disable blurscreen (Pull Request #2; thanks, @TechTrekkie!)
 - Updated `message` variable to clarify update instructions
 - Tweaked `updateScriptLog` function to satisfy my CDO (i.e., the alphabetical version of "OCD")